skycam/packages/skycam-common/pyproject.toml
[project]
name = "skycam-common"
version = "0.1.0"
description = "Shared utilities for skycam tools."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gphoto2==2.6.3",
    "pyyaml"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov"
]
